feat: creating assignments erros validaiton

+++ src/modules/assignments/data/services/delete-assignmente.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { DeleteAssignment } from "../../domain/usecases";
 import { AssignmentRepository } from "../contracts";
+import { RegisterNotFound } from "src/modules/chore/errors";
 
 @Injectable()
 export class DeleteAssignmentService implements DeleteAssignment {
@@ -9,6 +10,12 @@ export class DeleteAssignmentService implements DeleteAssignment {
     ) {}
 
     async execute(assignmentId: string): Promise<void> {
+        const assignment = await this.assignmentRepository.getById(assignmentId);
+
+        if(!assignment) {
+            throw new RegisterNotFound('tarefa')
+        }
+
         await this.assignmentRepository.delete(assignmentId)
     }
 }
